refact : notifcation delete 로직 반환값 int 애서 long으로 수정

+++ src/main/java/com/umc/naoman/domain/notification/repository/NotificationRepository.java
@@ -17,10 +17,10 @@ public interface NotificationRepository extends JpaRepository<Notification,Long>
     List<Notification> findAllByMemberIdAndIsCheckedFalse(Long memberId);
     @Modifying
     @Query("DELETE FROM Notification n WHERE n.member.id = :memberId AND n.id = :notificationId")
-    int deleteByMemberIdAndNotificationId(@Param("memberId") Long memberId, @Param("notificationId") Long notificationId);
+    long deleteByMemberIdAndNotificationId(@Param("memberId") Long memberId, @Param("notificationId") Long notificationId);
 
     @Modifying
     @Query("DELETE FROM Notification n WHERE n.member.id = :memberId")
-    int deleteByMemberId(@Param("memberId") Long memberId);
+    long deleteByMemberId(@Param("memberId") Long memberId);
 
 }
